Update: Computer choice logic
Change logic that uses math.random generate the computer choice.

Whereas v1 <commented out in code> was "Pick a number between 0 and 1
and split the ranges", v2 is saying "Just roll a dice with 3 sides
and use the result". This makes it easier to update the code with more
choices later if needed.